domains/data-collectors/carbon-intensity-service/requirements-prod.txt
fastapi>=0.115.0,<1.0.0
uvicorn>=0.34.0,<1.0.0
aiohttp>=3.9.0,<4.0.0  # HTTP client for WattTime API
python-dotenv>=1.0.1,<2.0.0
influxdb3-python>=0.3.0,<1.0.0
pydantic-settings>=2.0.0,<3.0.0

# HomeIQ shared libraries (installed from libs/ in Docker)
homeiq-observability>=1.0.0
homeiq-data>=1.0.0
homeiq-resilience>=1.0.0
